ObjectandClass
**Description**
this code is about the Lo1 In our OOP.
**Implementation**
<img width="734" alt="image" src="https://github.com/user-attachments/assets/27c20618-10d6-4b8b-9c3a-c01b9c4110ce">
<img width="339" alt="image" src="https://github.com/user-attachments/assets/33e597cd-51f4-412d-a45f-b081995ecaa4">

**Conclusion**
This HTML code creates a simple web page displaying a MY personal details using JavaScript. 
It has also features a title and headings for my name, my age, and my occupation. The person's info is stored
in a JavaScript object and shown on the page and in the console, showcasing basic web programming concepts.