Create object and class
**Description**
this code is about the LO1 object and classes

**Implementation**



**Conslusion**
This HTML code creates a simple web page displaying a MY personal details using JavaScript. 
It has also features a title and headings for my name, my age, and my occupation. The person's info is stored
in a JavaScript object and shown on the page and in the console, showcasing basic web programming concepts.

This HTML code creates a simple web page displaying a MY personal details using JavaScript. 
It has also features a title and headings for my name, my age, and my occupation. The person's info is stored
in a JavaScript object and shown on the page and in the console, showcasing basic web programming concepts.